backend/pyproject.toml
[tool.poetry]
name = "Matchmove_machine_site"
version = "0.1.0"
description = "Описание"
authors = ["timofejkazakov <t.kazakov@agora.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
redis = "^5.0.5"
django-celery-beat = "^2.6.0"
eventlet = "^0.36.1"
paddle-python-sdk = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
